+++ shared/lib_geothermal.cpp
@@ -2,7 +2,7 @@
 *  Copyright 2017 Alliance for Sustainable Energy, LLC
 *
 *  NOTICE: This software was developed at least in part by Alliance for Sustainable Energy, LLC
-*  (“Alliance”) under Contract No. DE-AC36-08GO28308 with the U.S. Department of Energy and the U.S.
+*  (ï¿½Allianceï¿½) under Contract No. DE-AC36-08GO28308 with the U.S. Department of Energy and the U.S.
 *  The Government retains for itself and others acting on its behalf a nonexclusive, paid-up,
 *  irrevocable worldwide license in the software to reproduce, prepare derivative works, distribute
 *  copies to the public, perform publicly and display publicly, and to permit others to do so.
@@ -26,8 +26,8 @@
 *  4. Redistribution of this software, without modification, must refer to the software by the same
 *  designation. Redistribution of a modified version of this software (i) may not refer to the modified
 *  version by the same designation, or by any confusingly similar designation, and (ii) must refer to
-*  the underlying software originally provided by Alliance as “System Advisor Model” or “SAM”. Except
-*  to comply with the foregoing, the terms “System Advisor Model”, “SAM”, or any confusingly similar
+*  the underlying software originally provided by Alliance as ï¿½System Advisor Modelï¿½ or ï¿½SAMï¿½. Except
+*  to comply with the foregoing, the terms ï¿½System Advisor Modelï¿½, ï¿½SAMï¿½, or any confusingly similar
 *  designation may not be used to refer to any modified version of this software or any modified
 *  version of the underlying software originally provided by Alliance without the prior written consent
 *  of Alliance.
@@ -1650,7 +1650,7 @@ bool CGeothermalAnalyzer::ReadWeatherForTimeStep(bool bHourly, unsigned int time
 
 	// Not an hourly analysis, so calculate the monthly weather info
 	int month = (timeStep % 12) + 1;
-	double hours = util::hours_in_month(month);
+	size_t hours = util::hours_in_month(month);
 	if (hours==0)
 	{
 		ms_ErrorString = "util::hours_in_month returned zero for month =  " + util::to_string(month) + ".";
@@ -1658,7 +1658,7 @@ bool CGeothermalAnalyzer::ReadWeatherForTimeStep(bool bHourly, unsigned int time
 	}
 
 	double pressure=0, wetbulb=0, drybulb=0, rel_humidity=0;
-	for (int i = 0; i<hours; i++)
+	for (size_t i = 0; i<hours; i++)
 	{
 		ReadNextLineInWeatherFile();
 		pressure += m_wf.pres;
@@ -1894,7 +1894,7 @@ bool CGeothermalAnalyzer::RunAnalysis( bool (*update_function)(float, void*), vo
 			}//hours
 
 			mp_geo_out->maf_monthly_resource_temp[iElapsedMonths] = (float)md_WorkingTemperatureC;	// resource temperature for this month
-			iEvaluationsInMonth = (IsHourly()) ? util::hours_in_month(month) : 1;
+			iEvaluationsInMonth = (IsHourly()) ? (unsigned int)util::hours_in_month(month) : 1;
 			mp_geo_out->maf_monthly_power[iElapsedMonths] = fMonthlyPowerTotal/iEvaluationsInMonth;		// avg monthly power
 			mp_geo_out->maf_monthly_energy[iElapsedMonths] = fMonthlyPowerTotal*util::hours_in_month(month)/iEvaluationsInMonth;		// energy output in month (kWh)
